organize test files  under presenter package
