+++ app/src/test/java/org/mifos/mobile/presenters/AccountOverviewPresenterTest.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
 
@@ -9,10 +9,11 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mifos.mobile.FakeRemoteDataSource
+import org.mifos.mobile.R
 
 import org.mifos.mobile.api.DataManager
 import org.mifos.mobile.models.client.ClientAccounts
-import org.mifos.mobile.presenters.AccountOverviewPresenter
 import org.mifos.mobile.ui.views.AccountOverviewMvpView
 import org.mifos.mobile.util.RxSchedulersOverrideRule
 import org.mockito.Mock
