+++ app/src/test/java/org/mifos/mobile/presenters/BeneficiaryDetailPresenterTest.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
 
@@ -11,9 +11,9 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mifos.mobile.R
 
 import org.mifos.mobile.api.DataManager
-import org.mifos.mobile.presenters.BeneficiaryDetailPresenter
 import org.mifos.mobile.ui.views.BeneficiaryDetailView
 import org.mifos.mobile.util.RxSchedulersOverrideRule
 import org.mockito.Mock
