+++ app/src/test/java/org/mifos/mobile/presenters/BeneficiaryListPresenterTest.kt
@@ -1,7 +1,6 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
-
 import io.reactivex.Observable
 
 import org.junit.After
@@ -9,12 +8,12 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mifos.mobile.FakeRemoteDataSource
+import org.mifos.mobile.R
 
 import org.mifos.mobile.api.DataManager
 import org.mifos.mobile.models.beneficiary.Beneficiary
-import org.mifos.mobile.models.templates.account.AccountOptionsTemplate
-import org.mifos.mobile.presenters.ThirdPartyTransferPresenter
-import org.mifos.mobile.ui.views.ThirdPartyTransferView
+import org.mifos.mobile.ui.views.BeneficiariesView
 import org.mifos.mobile.util.RxSchedulersOverrideRule
 
 import org.mockito.Mock
@@ -22,10 +21,10 @@ import org.mockito.Mockito
 import org.mockito.junit.MockitoJUnitRunner
 
 /**
- * Created by dilpreet on 24/7/17.
+ * Created by dilpreet on 27/6/17.
  */
 @RunWith(MockitoJUnitRunner::class)
-class ThirdPartyTransferPresenterTest {
+class BeneficiaryListPresenterTest {
     @Rule
     @JvmField
     val mOverrideSchedulersRule = RxSchedulersOverrideRule()
@@ -37,47 +36,43 @@ class ThirdPartyTransferPresenterTest {
     var dataManager: DataManager? = null
 
     @Mock
-    var view: ThirdPartyTransferView? = null
-    private var accountOptionsTemplate: AccountOptionsTemplate? = null
-    private var presenter: ThirdPartyTransferPresenter? = null
+    var view: BeneficiariesView? = null
+    private var presenter: BeneficiaryListPresenter? = null
     private var beneficiaryList: List<Beneficiary?>? = null
 
     @Before
+    @Throws(Exception::class)
     fun setUp() {
-        presenter = ThirdPartyTransferPresenter(dataManager!!, context!!)
+        presenter = BeneficiaryListPresenter(dataManager!!, context!!)
         presenter?.attachView(view)
-        accountOptionsTemplate = FakeRemoteDataSource.accountOptionsTemplate
         beneficiaryList = FakeRemoteDataSource.beneficiaries
     }
 
     @After
+    @Throws(Exception::class)
     fun tearDown() {
         presenter?.detachView()
     }
 
     @Test
-    fun testTransferTemplate() {
-        Mockito.`when`(dataManager?.thirdPartyTransferTemplate).thenReturn(Observable.just(accountOptionsTemplate))
+    @Throws(Exception::class)
+    fun testLoadBeneficiaries() {
         Mockito.`when`(dataManager?.beneficiaryList).thenReturn(Observable.just(beneficiaryList))
-        presenter?.loadTransferTemplate()
+        presenter?.loadBeneficiaries()
         Mockito.verify(view)?.showProgress()
         Mockito.verify(view)?.hideProgress()
-        Mockito.verify(view)?.showThirdPartyTransferTemplate(accountOptionsTemplate)
         Mockito.verify(view)?.showBeneficiaryList(beneficiaryList)
-        Mockito.verify(view, Mockito.never())?.showError(context?.getString(
-                R.string.error_fetching_account_transfer_template))
+        Mockito.verify(view, Mockito.never())?.showError(context?.getString(R.string.error_fetching_beneficiaries))
     }
 
     @Test
-    fun testTransferTemplateFails() {
-        Mockito.`when`(dataManager?.thirdPartyTransferTemplate).thenReturn(Observable.error(RuntimeException()))
+    @Throws(Exception::class)
+    fun testLoadBeneficiariesFails() {
         Mockito.`when`(dataManager?.beneficiaryList).thenReturn(Observable.error(RuntimeException()))
-        presenter?.loadTransferTemplate()
+        presenter?.loadBeneficiaries()
         Mockito.verify(view)?.showProgress()
         Mockito.verify(view)?.hideProgress()
-        Mockito.verify(view)?.showError(context?.getString(
-                R.string.error_fetching_account_transfer_template))
-        Mockito.verify(view, Mockito.never())?.showThirdPartyTransferTemplate(accountOptionsTemplate)
+        Mockito.verify(view)?.showError(context?.getString(R.string.error_fetching_beneficiaries))
         Mockito.verify(view, Mockito.never())?.showBeneficiaryList(beneficiaryList)
     }
 }
