+++ app/src/test/java/org/mifos/mobile/presenters/ClientChargePresenterTest.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
 
@@ -9,11 +9,12 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mifos.mobile.FakeRemoteDataSource
+import org.mifos.mobile.R
 
 import org.mifos.mobile.api.DataManager
 import org.mifos.mobile.models.Charge
 import org.mifos.mobile.models.Page
-import org.mifos.mobile.presenters.ClientChargePresenter
 import org.mifos.mobile.ui.views.ClientChargeView
 import org.mifos.mobile.util.RxSchedulersOverrideRule
