+++ app/src/test/java/org/mifos/mobile/presenters/GuarantorDetailPresenterTest.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
