+++ app/src/test/java/org/mifos/mobile/presenters/GuarantorListPresenterTest.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
 
@@ -9,10 +9,10 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mifos.mobile.FakeRemoteDataSource
 
 import org.mifos.mobile.api.DataManager
 import org.mifos.mobile.models.guarantor.GuarantorPayload
-import org.mifos.mobile.presenters.GuarantorListPresenter
 import org.mifos.mobile.ui.views.GuarantorListView
 import org.mifos.mobile.util.RxSchedulersOverrideRule
