+++ app/src/test/java/org/mifos/mobile/presenters/LoanAccountWithdrawPresenterTest.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
 
@@ -11,10 +11,10 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mifos.mobile.R
 
 import org.mifos.mobile.api.DataManager
 import org.mifos.mobile.models.accounts.loan.LoanWithdraw
-import org.mifos.mobile.presenters.LoanAccountWithdrawPresenter
 import org.mifos.mobile.ui.views.LoanAccountWithdrawView
 import org.mifos.mobile.util.RxSchedulersOverrideRule
