+++ app/src/test/java/org/mifos/mobile/presenters/LoginPresenterTest.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
 
@@ -9,6 +9,9 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mifos.mobile.FakeRemoteDataSource
+import org.mifos.mobile.R
+import org.mifos.mobile.RetrofitUtils
 
 import org.mifos.mobile.api.BaseURL
 import org.mifos.mobile.api.DataManager
@@ -17,7 +20,6 @@ import org.mifos.mobile.models.Page
 import org.mifos.mobile.models.User
 import org.mifos.mobile.models.client.Client
 import org.mifos.mobile.models.payload.LoginPayload
-import org.mifos.mobile.presenters.LoginPresenter
 import org.mifos.mobile.ui.views.LoginView
 import org.mifos.mobile.util.RxSchedulersOverrideRule
 
@@ -98,7 +100,11 @@ class LoginPresenterTest {
     @Test
     @Throws(Exception::class)
     fun testLoadClientFails() {
-        Mockito.`when`(dataManager?.clients).thenReturn(Observable.error(RetrofitUtils.getResponseForError(404)))
+        Mockito.`when`(dataManager?.clients).thenReturn(Observable.error(
+            RetrofitUtils.getResponseForError(
+                404
+            )
+        ))
         presenter?.loadClient()
         Mockito.verify(view)?.hideProgress()
         Mockito.verify(view)?.showMessage(context?.getString(R.string.error_fetching_client))
@@ -108,7 +114,11 @@ class LoginPresenterTest {
     @Test
     @Throws(Exception::class)
     fun testLoadClientUnauthorized() {
-        Mockito.`when`(dataManager?.clients).thenReturn(Observable.error(RetrofitUtils.getResponseForError(401)))
+        Mockito.`when`(dataManager?.clients).thenReturn(Observable.error(
+            RetrofitUtils.getResponseForError(
+                401
+            )
+        ))
         presenter?.loadClient()
         Mockito.verify(view)?.hideProgress()
         Mockito.verify(view)?.showMessage(context?.getString(R.string.unauthorized_client))
