+++ app/src/test/java/org/mifos/mobile/presenters/RecentTransactionsPresenterTest.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
 
@@ -9,11 +9,12 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mifos.mobile.FakeRemoteDataSource
+import org.mifos.mobile.R
 
 import org.mifos.mobile.api.DataManager
 import org.mifos.mobile.models.Page
 import org.mifos.mobile.models.Transaction
-import org.mifos.mobile.presenters.RecentTransactionsPresenter
 import org.mifos.mobile.ui.views.RecentTransactionsView
 import org.mifos.mobile.util.RxSchedulersOverrideRule
 
@@ -64,7 +65,9 @@ class RecentTransactionsPresenterTest {
         Mockito.verify(view)?.showProgress()
         Mockito.verify(view)?.hideProgress()
         Mockito.verify(view)?.showRecentTransactions(recentTransaction?.pageItems)
-        Mockito.verify(view, Mockito.never())?.showErrorFetchingRecentTransactions(context?.getString(R.string.error_recent_transactions_loading))
+        Mockito.verify(view, Mockito.never())?.showErrorFetchingRecentTransactions(context?.getString(
+            R.string.error_recent_transactions_loading
+        ))
     }
 
     @Test
@@ -76,7 +79,9 @@ class RecentTransactionsPresenterTest {
         Mockito.verify(view)?.showProgress()
         Mockito.verify(view)?.hideProgress()
         Mockito.verify(view)?.showLoadMoreRecentTransactions(recentTransaction?.pageItems)
-        Mockito.verify(view, Mockito.never())?.showErrorFetchingRecentTransactions(context?.getString(R.string.error_recent_transactions_loading))
+        Mockito.verify(view, Mockito.never())?.showErrorFetchingRecentTransactions(context?.getString(
+            R.string.error_recent_transactions_loading
+        ))
     }
 
     @Test
@@ -88,7 +93,9 @@ class RecentTransactionsPresenterTest {
         Mockito.verify(view)?.showProgress()
         Mockito.verify(view)?.hideProgress()
         Mockito.verify(view)?.showEmptyTransaction()
-        Mockito.verify(view, Mockito.never())?.showErrorFetchingRecentTransactions(context?.getString(R.string.error_recent_transactions_loading))
+        Mockito.verify(view, Mockito.never())?.showErrorFetchingRecentTransactions(context?.getString(
+            R.string.error_recent_transactions_loading
+        ))
     }
 
     @Test
