+++ app/src/test/java/org/mifos/mobile/presenters/UpdatePasswordPresenterTest.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
 
@@ -12,12 +12,12 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mifos.mobile.FakeRemoteDataSource
 
 import org.mifos.mobile.api.BaseURL
 import org.mifos.mobile.api.DataManager
 import org.mifos.mobile.api.local.PreferencesHelper
 import org.mifos.mobile.models.UpdatePasswordPayload
-import org.mifos.mobile.presenters.UpdatePasswordPresenter
 import org.mifos.mobile.ui.views.UpdatePasswordView
 import org.mifos.mobile.util.RxSchedulersOverrideRule
 import org.mifos.mobile.utils.MFErrorParser.errorMessage
