+++ app/src/test/java/org/mifos/mobile/presenters/UserDetailsPresenterTest.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile
+package org.mifos.mobile.presenters
 
 import android.content.Context
 
@@ -11,12 +11,13 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mifos.mobile.FakeRemoteDataSource
+import org.mifos.mobile.R
 
 import org.mifos.mobile.api.DataManager
 import org.mifos.mobile.api.local.PreferencesHelper
 import org.mifos.mobile.models.client.Client
 import org.mifos.mobile.models.notification.NotificationRegisterPayload
-import org.mifos.mobile.presenters.UserDetailsPresenter
 import org.mifos.mobile.ui.views.UserDetailsView
 import org.mifos.mobile.util.RxSchedulersOverrideRule
